hidden quest #3 - 9461 2nd try: success (fix overflow)

+++ hidden_quest#3/04_9461.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int T, N;
 
-int dp[101];
+long long dp[101];
 int main () {
   cin >> T;
